Clean up server.js
Split up the server into multiple routes for better readability
